# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cyrano api swagger",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/projects": {
            "post": {
                "summary": "Create a Project",
                "description": "Create a new Project",
                "operationId": "project.create",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/project"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/channels/poynts": {
            "post": {
                "summary": "Create a new Poynt Channel",
                "description": "Create a new Poynt Channel",
                "operationId": "project.channels.poynt.create",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "description": "ID of the project",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/poynt_channel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "404": {
                        "description": "project not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/v1/projects/channels/poynts": {
            "get": {
                "summary": "Find a Poynt Project and Channel by business_id and device_id",
                "description": "Find a Poynt Project and Channel by business_id and device_id",
                "operationId": "project.channels.poynt.find",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "description": "Business ID",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "description": "Device ID",
                        "required": true,
                        "type": "string",
                        "format": "urn:tid:uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "not acceptable"
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "403": {
                        "description": "another channel is already associated"
                    },
                    "404": {
                        "description": "channel not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "project": {
            "description": "Project",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Project name",
                    "type": "string",
                    "example": "Default Poynt Project"
                }
            },
            "type": "object"
        },
        "poynt_channel": {
            "description": "Poynt Channel",
            "required": [
                "name",
                "business_id",
                "device_id"
            ],
            "properties": {
                "name": {
                    "description": "Channel name",
                    "type": "string",
                    "example": "Default Poynt Channel"
                },
                "business_id": {
                    "description": "Business ID",
                    "type": "string",
                    "format": "uuid",
                    "example": "18f071cc-5ed4-4b33-80c1-305056d42bfb"
                },
                "device_id": {
                    "description": "Device ID",
                    "type": "string",
                    "format": "urn:tid:uuid",
                    "example": "urn:tid:48c54303-6d51-39af-bdeb-4af53f621652"
                }
            },
            "type": "object"
        }
    }
}